Split typechecking of toplevel declarations
