+++ lang/elaborator/src/typechecker/decls/definition.rs
@@ -1,0 +1,44 @@
+use syntax::ast::*;
+
+use crate::normalizer::{env::ToEnv, normalize::Normalize};
+
+use crate::typechecker::{
+    ctx::Ctx,
+    typecheck::{CheckInfer, InferTelescope, WithScrutinee},
+    util::ExpectTypApp,
+    TypeError,
+};
+
+use super::CheckToplevel;
+
+/// Infer a definition
+impl CheckToplevel for Def {
+    fn check_wf(&self, prg: &Module, ctx: &mut Ctx) -> Result<Self, TypeError> {
+        let Def { span, doc, name, attr, params, self_param, ret_typ, body } = self;
+
+        params.infer_telescope(prg, ctx, |ctx, params_out| {
+            let self_param_nf = self_param.typ.normalize(prg, &mut ctx.env())?;
+
+            let (ret_typ_out, ret_typ_nf, self_param_out) =
+                self_param.infer_telescope(prg, ctx, |ctx, self_param_out| {
+                    let ret_typ_out = ret_typ.infer(prg, ctx)?;
+                    let ret_typ_nf = ret_typ.normalize(prg, &mut ctx.env())?;
+                    Ok((ret_typ_out, ret_typ_nf, self_param_out))
+                })?;
+
+            let body_out =
+                WithScrutinee { inner: body, scrutinee: self_param_nf.expect_typ_app()? }
+                    .check_ws(prg, ctx, ret_typ_nf)?;
+            Ok(Def {
+                span: *span,
+                doc: doc.clone(),
+                name: name.clone(),
+                attr: attr.clone(),
+                params: params_out,
+                self_param: self_param_out,
+                ret_typ: ret_typ_out,
+                body: body_out,
+            })
+        })
+    }
+}
